feat(expandable): Update chevrons and documentation
