+++ modules/labs-react/expandable/README.md
@@ -4,97 +4,8 @@
   <img src="https://img.shields.io/badge/LABS-alpha-orange" alt="LABS: Alpha" />
 </a>  This component is work in progress and currently in prerelease.
 
-Container that expands content
-
-## Installation
-
-```sh
-yarn add @workday/canvas-kit-labs-react
-```
-
 ## Usage
 
-### Basic Example
-
-```tsx
-import * as React from 'react';
-import {Expandable} from '@workday/canvas-kit-labs-react/expandable';
-
-<Expandable>
-  <Expandable.Target headingLevel="h1">
-    <Expandable.StartChevron />
-    Additional Information
-  </Expandable.Target>
-  <Expandable.Content>Content</Expandable.Content>
-</Expandable>;
-```
-
-## Components
-
-### Expandable
-
-#### Component Props
-
-##### Required
-
-> children
-
-##### Optional
-
-> model
-
-### Expandable.Target
-
-#### Component Props
-
-##### Required
-
-> headingLevel children
-
-##### Optional
-
-> model
-
-### Expandable.Title
-
-#### Component Props
-
-##### Required
-
-> children
-
-### Expandable.StartChevron
-
-#### Component Props
-
-##### Optional
-
-> model
-
-### Expandable.EndChevron
-
-#### Component Props
-
-##### Optional
-
-> model
-
-### Expandable.Avatar
-
-#### Component Props
-
-##### Optional
-
-> AvatarProps
-
-### Expandable.Content
-
-#### Component Props
-
-##### Required
-
-> children
-
-##### Optional
-
-> model
+View the
+[documentation for Expandable](https://workday.github.io/canvas-kit/?path=/docs/labs-expandable-react--basic)
+on Storybook.
