+++ modules/labs-react/expandable/stories/examples/Avatar.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+import {Expandable} from '@workday/canvas-kit-labs-react/expandable';
+
+export const Avatar = () => (
+  <Expandable>
+    <Expandable.Target headingLevel="h3">
+      <Expandable.StartChevron />
+      <Expandable.Title>
+        <Expandable.Avatar />
+        Title
+      </Expandable.Title>
+    </Expandable.Target>
+
+    <Expandable.Content>Content</Expandable.Content>
+  </Expandable>
+);
